(#34) Move purely UI parts into a separate mod
