+++ src/ui/keycodes.rs
@@ -1,0 +1,8 @@
+pub const KEY_E      : i32 = 0x65;
+pub const KEY_Q      : i32 = 0x71;
+pub const KEY_TAB    : i32 = 0x09;
+pub const KEY_RETURN : i32 = 0x0a;
+pub const KEY_S      : i32 = 0x73;
+pub const KEY_W      : i32 = 0x77;
+pub const KEY_D      : i32 = 0x64;
+pub const KEY_A      : i32 = 0x61;
